Fix error when using exe wavtool on linux
The system for running exe wavtools is to create a temporary batch
file, but OpenUtau tries to run this batch file directly, which
does not work on Linux. This commit adds a check for the operating
system and will create a shell script wrapper that runs the batch
file through wine if the OS is Linux. A few other changes also had
to be made for this to work, such as using different character
encodings that did not include byte order marks, because they
cause problems both when running shell scripts and when running
batch files through wine. Because the current implementation uses
the path style appropriate for the operating system, it was
writing linux-style paths to the batch file, which was another
cause of issues that this commit fixes. Finally, the wavtool and
resampler paths both point to the wrapper created by the user
during the install process (see link at bottom), but this is not
the correct path to give the batch file because windows batch
files cannot run ilnux shell scripts; therefore, this commit
includes code to extract the path to the exe file from the sh
wrapper.

https://github.com/stakira/OpenUtau/wiki/Resamplers-and-Wavtools#linux