Audio Export Configuration Enhancements
A new constructor has been added to the `ExportAdapter` that allows the number of channels and sampling rate to be specified. With this change, the export process in `PlaybackManager` has been updated and the call to the `ToMono` method has been removed.

Added properties for number of channels and sampling rate to the user preferences, with default values. The settings screen now has a combo box to select these new options, and the user interface has been improved.