Added description of class `F0Smoother` and removed SmoothPitch
A detailed description of the class `F0Smoother` has been added to `F0Smoother.cs`. This includes the purpose of the class, a link to the original Python implementation and copyright information.

Removed the `SmoothPitch` method from `VoicevoxUtils.cs`. This method had the ability to smooth the pitch and expressivity arrays, but was removed as it was no longer needed.