feat(onnx): drop Vortice.DXGI & update dml to 1.23.0
> **Note**: This change requires an update on package `Microsoft.ML.OnnxRuntime.DirectML` from `1.16.3` to `1.23.0`.

In some platforms with `Intel(R) AI Boost` NPU, DXGI will report it as a `Microsoft Basic Render Driver` and this will cause `onnxruntime` panic because it actually do not support it.

In order to eliminate this issue, the native `GetEpDevices` from package `Microsoft.ML.OnnxRuntime` is called, which will replace DXGI and makes it possible to be extended to other device types (like CoreML, OpenVINO, etc.) more easily in the future.

I have tested a DiffSinger model locally and currently no error has been found.