Adding the OverwriteLinePitchTool and Implementing Related Features
Added the `OverwriteLinePitchTool` property to `NotesViewModel` and initialized it in the constructor. Also reflected this property's state in `SelectToolCommand`.

Added `OverwriteLinePitchState` to the `NoteEditStates` class and implemented functionality to override specific pitches.

In the UI, a toggle button was added to `PianoRollWindow.axaml`. The conditions in the `NotesCanvasLeftPointerPressed` and `NotesCanvasRightPointerPressed` methods were updated to consider the state of the `OverwriteLinePitchTool`.

Additionally, the conditions in the `NotesCanvasPointerMoved` method were similarly updated.